Refactored methods from file generation into common module
