+++ lib/tecan_file_generation.rb
@@ -1,68 +1,30 @@
 module Sanger
   module Robots
     module Tecan
+      #
+      # Generates the picking file contents for the Tecan robot
+      #
       class Generator
         class << self
+          include ::CommonFileGenerator
+
           def mapping(data_object, total_volume)
-            raise ArgumentError, 'data_object needs to conform to an interface. WIP' if data_object.nil?
+            raise ArgumentError, 'Data object not present for Tecan mapping' if data_object.nil?
 
             dest_barcode_index = barcode_to_plate_index(data_object['destination'])
-
             source_barcode_index = source_barcode_to_plate_index(data_object['destination'])
-            buffer_data = buffers(data_object, total_volume)
+
             output_file_contents = [header(data_object)]
+
+            buffer_data = buffers(data_object, total_volume)
             if buffer_data.present?
               output_file_contents << buffer_data
               output_file_contents << buffer_seperator
             end
+
             output_file_contents << dyn_mappings(data_object)
             output_file_contents << footer(source_barcode_index, dest_barcode_index)
-
             output_file_contents.join("\n").gsub(/\n\n/, "\n")
-          end
-
-          def barcode_to_plate_index(plates)
-            barcode_lookup = {}
-            plates.each_with_index do |plate, index|
-              barcode_lookup[plate[0]] = index + 1
-            end
-            barcode_lookup
-          end
-
-          def source_barcode_to_plate_index(destination)
-            all_barcodes = []
-            barcode_lookup = {}
-            destination.each do |plate_id, plate_info|
-              # sort by destination well to make sure the plates are put the right way round for the robot
-              # 'SCRC1' goes into the 1st row of the fluidigm chip, and 'SCRC2' into the 2nd
-              mapping_sorted = sort_mapping_by_destination_well(plate_id, plate_info['mapping'])
-              mapping_sorted.each do |map_well|
-                well = map_well['src_well']
-                all_barcodes << well[0]
-              end
-            end
-            all_barcodes.uniq.each_with_index do |plate, index|
-              barcode_lookup[plate] = index + 1
-            end
-            barcode_lookup
-          end
-
-          def sort_mapping_by_destination_well(plate_id, mapping)
-            # query relevant 'map' records based on asset shape id & asset size, then sort by row order
-            # return the original mapping if the Plate cannot be found using the barcode - for instance, if this is coming from stock_stamper.rb
-            plate = Plate.find_by_barcode(plate_id)
-            return mapping if plate.nil?
-
-            purpose = Purpose.find(plate.plate_purpose_id)
-
-            relevant_map_records = Map.where(asset_shape_id: purpose.asset_shape_id, asset_size: plate.size)
-            relevant_map_records_by_description = {}
-            relevant_map_records.each { |map_record| relevant_map_records_by_description[map_record.description] = map_record }
-
-            mapping.sort_by do |a|
-              map_record_description = a['dst_well']
-              relevant_map_records_by_description[map_record_description].row_order
-            end
           end
 
           private
